net/ice/base: refactor (non) bitmap declarations
The firmware interface definition file mistakenly started using bitmap
declarations for its storage in a couple of structures. The data
structure contains a "bit-map" of bits that the hardware/firmware sets
or reads but it's not a DECLARE_BITMAP style bitmap and it's not
accessed with the set_bit/clear_bit operators, so stop declaring them
that way.

Signed-off-by: Jesse Brandeburg <jesse.brandeburg@intel.com>
Signed-off-by: Ian Stokes <ian.stokes@intel.com>
Acked-by: Bruce Richardson <bruce.richardson@intel.com>