net/ice/base: use bitmap type in promisc functions
Currently all functions related to promiscuous modes use u8 bit mask.
All 8 bits are utilized already and to be able to add any new flag, all
the drivers need to comply.

To address that, change u8 into ice_bitmap_t and use dedicated functions
to process it.

Updates made to DPDK-specific driver files also to align with this
change in the base code.

Signed-off-by: Mateusz Pacuszka <mateuszx.pacuszka@intel.com>
Signed-off-by: Ian Stokes <ian.stokes@intel.com>
Acked-by: Bruce Richardson <bruce.richardson@intel.com>