net/ice/base: enable RDMA Act-Act unload paths
In certain unload conditions, the non-standard paths to breaking down the
bond-supporting resources need special considerations in that there
could be nodes moved out of the tree. This can cause a NULL pointer
to be passed into the find node function.

Add in a NULL pointer check to handle these situations

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>
Signed-off-by: Ian Stokes <ian.stokes@intel.com>
Acked-by: Bruce Richardson <bruce.richardson@intel.com>